+++ docs/changelog.rst
@@ -9,6 +9,15 @@ The format is based on `Keep a Changelog`_, and this project adheres to
 Unreleased
 ----------
 
+[0.6.4] 2023-08-22
+------------------
+
+Fixed
+~~~~~
+
+- Bug in :func:`xugrid.snap_to_grid`, which caused an ``IndexError``. 
+  See `#122 <https://github.com/Deltares/xugrid/issues/122>`_.
+
 
 [0.6.3] 2023-08-12
 ------------------
@@ -36,7 +45,7 @@ Added
 - :meth:`xugrid.UgridDatasetAccessor.rename` and
   :meth:`xugrid.UgridDataArrayAccessor.rename` to rename both topology and the
   associated dimensions.
-- :meth:`xugrid.Ugrid2d.polygon_bounds` has been added to get a polygon
+- :meth:`xugrid.Ugrid2d.bounding_polygon` has been added to get a polygon
   describing the bounds of the grid.
 
 Fixed